Refactored non UI components into their own libraries and converted classes to file based namespaces
