+++ src/EventLogExpert.Store/Actions/FilterPaneAction.cs
@@ -1,0 +1,6 @@
+﻿namespace EventLogExpert.Store.Actions;
+
+public record FilterPaneAction
+{
+    public record AddRecentFilter(string FilterText);
+}
