+++ src/EventLogExpert.Store/Actions/StatusBarAction.cs
@@ -1,0 +1,6 @@
+﻿namespace EventLogExpert.Store.Actions;
+
+public record StatusBarAction
+{
+    public record SetEventsLoaded(int EventCount);
+}
