+++ src/EventLogExpert/App.xaml.cs
@@ -2,10 +2,10 @@
 
 public partial class App : Application
 {
-	public App()
-	{
-		InitializeComponent();
+    public App()
+    {
+        InitializeComponent();
 
-		MainPage = new NavigationPage(new MainPage());
-	}
+        MainPage = new NavigationPage(new MainPage());
+    }
 }
