+++ src/EventLogExpert/Components/StatusBar.razor
@@ -1,17 +1,4 @@
-﻿@inherits Fluxor.Blazor.Web.Components.FluxorComponent
-@using EventLogExpert.Store
-@using Fluxor
-@inject IState<StatusBarState> statusBarState
+﻿@inherits FluxorComponent
+@inject IState<StatusBarState> StatusBarState
 
-<style>
-    .status-bar {
-        background-color: #3C64E7;
-        color: white;
-    }
-</style>
-
-<div class="status-bar">Events Loaded: @statusBarState.Value.EventsLoaded</div>
-
-@code {
-
-}
+<div class="status-bar">Events Loaded: @StatusBarState.Value.EventsLoaded</div>
