--- src/EventLogExpert/EventUtils/NativeMethods.cs
@@ -1,37 +1,0 @@
-﻿using System;
-using System.Runtime.InteropServices;
-
-// ReSharper disable InconsistentNaming
-// We are defining some win32 types in this file, so we
-// are not following the usual C# naming conventions.
-
-namespace EventLogExpert.EventUtils
-{
-    public class NativeMethods
-    {
-        public const int RT_MESSAGETABLE = 11;
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        public static extern IntPtr LoadLibrary(string fileName);
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        public static extern bool FreeLibrary(IntPtr hModule);
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        public static extern IntPtr FindResource(IntPtr hModule, int lpID, int lpType);
-
-        [DllImport("kernel32.dll", SetLastError = true)]
-        public static extern IntPtr LoadResource(IntPtr hModule, IntPtr hResInfo);
-
-        [DllImport("kernel32.dll")]
-        public static extern IntPtr LockResource(IntPtr hResData);
-
-        [StructLayout(LayoutKind.Sequential)]
-        public struct MESSAGE_RESOURCE_BLOCK
-        {
-            public int LowId;
-            public int HighId;
-            public int OffsetToEntries;
-        }
-    }
-}
