+++ src/EventLogExpert/_Imports.razor
@@ -6,3 +6,8 @@
 @using Microsoft.JSInterop
 @using EventLogExpert
 @using EventLogExpert.Shared
+@using EventLogExpert.Store
+@using EventLogExpert.Store.State
+
+@using Fluxor
+@using Fluxor.Blazor.Web.Components
