+++ src/EventLogExpert/wwwroot/index.html
@@ -23,6 +23,7 @@
 	</div>
 
 	<script src="_framework/blazor.webview.js" autostart="false"></script>
+    <script src="_content/Fluxor.Blazor.Web/scripts/index.js"></script>
 
 </body>
